Rename Dashboard styledComponents and textComponents files
